stock_market/stock_market/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
yfinance==0.2.28
pandas==2.1.4
numpy==1.24.4
scikit-learn==1.3.2
python-dateutil==2.8.2
pydantic==2.5.1
python-multipart==0.0.6
httpx==0.25.2
python-dotenv==1.0.0
urllib3==2.2.3
requests==2.32.3
